FU-129 refactor: move customer side info containers
